% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_boem_wea_outlines.R
\docType{data}
\name{boem_wea_outlines}
\alias{boem_wea_outlines}
\title{BOEM Renewable Energy Lease Areas and Wind Planning Areas}
\format{
A \code{sf} collection with 53 features and 23 fields:
\describe{
\item{Geometry type}{MULTIPOLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -158.6636 ymin: 20.77651 xmax: -70.02155 ymax: 43.95971}
\item{Geodetic CRS}{WGS 84}
\item{source}{data-raw/boem-renewable-energy-geodatabase/BOEMWindLayers_4Download.gdb}
}
}
\source{
\url{https://www.boem.gov/renewable-energy/mapping-and-data/renewable-energy-gis-data}
}
\usage{
data("boem_wea_outlines")
}
\description{
A \code{sf} object containing the outlines for BOEM Renewable Energy Lease Areas (LEASE_STAGE = 'Active') and Wind Planning Areas (LEASE_STAGE = 'Planning').
}
\keyword{datasets}
