% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-coastspan_strata.R
\docType{data}
\name{coastspan_strata}
\alias{coastspan_strata}
\title{Cooperative Atlantic States Shark Pupping and Nursery Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -81.6 ymin: 27.1 xmax: -74.9 ymax: 39.2}
\item{Projected CRS}{WGS 84}
\item{Features}{7 x 5}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('coastspan_strata')
}
\description{
An \code{sf} object containing spatial data for the Cooperative Atlantic States Shark Pupping and Nursery Survey.
}
\examples{
\dontrun{
# Load the dataset
data(coastspan_strata)

# Plot the geometry
plot(sf::st_geometry(coastspan_strata))
}
}
\keyword{{datasets}}
