% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-edna_stations.R
\docType{data}
\name{edna_stations}
\alias{edna_stations}
\title{eDNA Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POINT}
\item{Dimension}{NA}
\item{Bounding box}{xmin: -75.5 ymin: 36 xmax: -65.4 ymax: 44.5}
\item{Projected CRS}{WGS 84}
\item{Features}{35 x 13}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('edna_stations')
}
\description{
An \code{sf} object containing spatial data for the eDNA Survey. Sampling stations for the eDNA Survey.
}
\examples{
\dontrun{
# Load the dataset
data(edna_stations)

# Plot the geometry
plot(sf::st_geometry(edna_stations))
}
}
\keyword{{datasets}}
