% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_final_mass_inshore_stat.R
\docType{data}
\name{final_mass_inshore_stat}
\alias{final_mass_inshore_stat}
\title{Massachusetts Inshore Statistical Areas}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{MULTIPOLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -71.12554 ymin: 41.18887 xmax: -69.86013 ymax: 42.87647}
\item{Geodetic CRS}{NAD83}
\item{Features}{14 x 5}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.fisheries.noaa.gov/inport/item/33467}
}
\usage{
data("final_mass_inshore_stat")
}
\description{
An \code{sf} object containing the boundaries of the Massachusetts Inshore statistical areas.
}
\examples{
\dontrun{
# Load the dataset
data(final_mass_inshore_stat)

# Plot the geometry
plot(sf::st_geometry(final_mass_inshore_stat))
}
}
\keyword{datasets}
