% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mdeb_data.R
\name{find_mdeb_data}
\alias{find_mdeb_data}
\title{Find available data hosted on the MDEB (Marine Development and Ecology Branch) GIS Data Hub
hosted on NOAA's Geoplatform}
\usage{
find_mdeb_data()
}
\value{
A tibble with the following columns:
\item{title}{Name of the NOAA Geoplatform feature service}
\item{name}{Name of the hosted feature layer in the feature service}
\item{id}{Feature layer ID, a unique identifier assigned to layers within a feature service}
\item{url}{URL of the NOAA Geoplatform feature layer}
}
\description{
This function retrieves the data available on the
\href{https://mdeb-nefsc-noaa.hub.arcgis.com/}{MDEB GIS Data Hub}.
}
\examples{
# Retrieve a list of available data
mdeb_hub_datasets <- find_mdeb_datasets()
mdeb_hub_datasets 

}
