% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_FootprintAreaProjected.R
\docType{data}
\name{FootprintAreaProjected}
\alias{FootprintAreaProjected}
\title{Northwest Atlantic Fisheries Organization (NAFO) footprint}
\format{
A data frame/ sf object (1 x 6)
\describe{
\item{Id}{}
\item{ET_ID}{}
\item{Area}{km^2}
\item{Perimeter}{in km}
\item{X}{x coordinate of polygon center}
\item{y}{y coordinate of polygon center}
}
}
\source{
\url{https://www.nafo.int/Data/GIS}
}
\usage{
FootprintAreaProjected
}
\description{
An sf object containing the NAFO footprint
Pulled from https://www.nafo.int/Data/GIS
}
\section{Projection}{


CRS: NAD83/ UTM Zone 23N
}

\seealso{
Other NAFO: 
\code{\link{NAFO_Divisions_2021_line_enclosed}},
\code{\link{NAFO_Divisions_2021_line_not_enclosed}},
\code{\link{NAFO_Divisions_2021_poly_clipped}},
\code{\link{NAFO_Divisions_2021_poly_not_clipped}},
\code{\link{NAFO_Divisions_2021_pts}}
}
\concept{NAFO}
\keyword{datasets}
