% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_footprint.R
\name{get_stock_footprint}
\alias{get_stock_footprint}
\title{create species stock area}
\usage{
get_stock_footprint(itis_code, stock_def)
}
\arguments{
\item{itis_code}{Species itis code for species}

\item{stock_def}{Name of the stock area, eg. SNEMA, UNIT, NORTH etc}
}
\value{
sf object
}
\description{
Allows users to filter the Bottom trawl survey sf object by strata for many species
with defined stock areas. Stock definitions are found in the lazily loaded
data: \code{stock_footprints}
}
