% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hl_strata.R
\docType{data}
\name{hl_strata}
\alias{hl_strata}
\title{Hook and Line Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -75.9 ymin: 35.3 xmax: -68.6 ymax: 41.8}
\item{Projected CRS}{WGS 84}
\item{Features}{1 x 6}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('hl_strata')
}
\description{
An \code{sf} object containing spatial data for the Hook and Line Survey.
}
\examples{
\dontrun{
# Load the dataset
data(hl_strata)

# Plot the geometry
plot(sf::st_geometry(hl_strata))
}
}
\keyword{{datasets}}
