% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mmst_strata.R
\docType{data}
\name{mmst_strata}
\alias{mmst_strata}
\title{Marine Mammal and Sea Turtle Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -82.1 ymin: 24.2 xmax: -61.8 ymax: 45.2}
\item{Projected CRS}{WGS 84}
\item{Features}{7 x 6}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('mmst_strata')
}
\description{
An \code{sf} object containing spatial data for the Marine Mammal and Sea Turtle Survey.
}
\examples{
\dontrun{
# Load the dataset
data(mmst_strata)

# Plot the geometry
plot(sf::st_geometry(mmst_strata))
}
}
\keyword{{datasets}}
