% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_NAFO_Divisions_2021_line_enclosed.R
\docType{data}
\name{NAFO_Divisions_2021_line_enclosed}
\alias{NAFO_Divisions_2021_line_enclosed}
\title{NAFO Divisions (Bounded)}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{LINESTRING}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -80 ymin: 35 xmax: -42 ymax: 78.16667}
\item{Geodetic CRS}{NAD83}
\item{Features}{38 x 6}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.nafo.int/Data/GIS}
}
\usage{
data("NAFO_Divisions_2021_line_enclosed")
}
\description{
An \code{sf} object that depicts the boundaries of the NAFO regions
}
\examples{
\dontrun{
# Load the dataset
data(NAFO_Divisions_2021_line_enclosed)

# Plot the geometry
plot(sf::st_geometry(NAFO_Divisions_2021_line_enclosed))
}
}
\keyword{datasets}
