% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_NAFO_Divisions_2021_poly_clipped.R
\docType{data}
\name{NAFO_Divisions_2021_poly_clipped}
\alias{NAFO_Divisions_2021_poly_clipped}
\title{NAFO Divisions (Polygon clipped to coastline)}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -80 ymin: 35 xmax: -42 ymax: 78.16667}
\item{Geodetic CRS}{NAD83}
\item{Features}{39 x 9}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.nafo.int/Data/GIS}
}
\usage{
data("NAFO_Divisions_2021_poly_clipped")
}
\description{
An \code{sf} object that depicts the boundaries of the NAFO regions.
}
\examples{
\dontrun{
# Load the dataset
data(NAFO_Divisions_2021_poly_clipped)

# Plot the geometry
plot(sf::st_geometry(NAFO_Divisions_2021_poly_clipped))
}
}
\keyword{datasets}
