% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEFSCspatial-package.R
\docType{package}
\name{NEFSCspatial-package}
\alias{NEFSCspatial}
\alias{NEFSCspatial-package}
\title{NEFSCspatial: Spatial data package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides spatial data as sf objects
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NEFSC/NEFSC-Spatial}
  \item Report bugs at \url{https://github.com/NEFSC/NEFSC-Spatial/issues}
}

}
\author{
\strong{Maintainer}: Andy Beet \email{andrew.beet@noaa.gov}

Authors:
\itemize{
  \item Kimberly Bastille \email{kimberly.bastille@noaa.gov}
  \item Joshua Hatch \email{joshua.hatch@noaa.gov}
}

}
\keyword{internal}
