% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Neus_atlantis.R
\docType{data}
\name{Neus_atlantis}
\alias{Neus_atlantis}
\title{Northeast US (NEUS) atlantis model domain}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -76.45919 ymin: 34.76128 xmax: -63.42171 ymax: 46.11478}
\item{Geodetic CRS}{WGS 83}
\item{Features}{30 x 11}
\item{Metadata}{See source}
}
}
\source{
\url{https://github.com/NOAA-EDAB/neus-atlantis/wiki/Spatial-extent-of-model}
}
\usage{
data("Neus_atlantis")
}
\description{
An sf object containing the box level boundaries used in the Northeast Atlantis ecosystem end to end model
}
\keyword{datasets}
