% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sc_strata.R
\docType{data}
\name{sc_strata}
\alias{sc_strata}
\title{Atlantic Surfclam and Ocean Quahog Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{MULTIPOLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -76 ymin: 36.5 xmax: -66.9 ymax: 42.1}
\item{Projected CRS}{WGS 84}
\item{Features}{12 x 7}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('sc_strata')
}
\description{
An \code{sf} object containing spatial data for the Atlantic Surfclam and Ocean Quahog Survey. Scientific survey strata for the Atlantic Surfclam Survey.
}
\examples{
\dontrun{
# Load the dataset
data(sc_strata)

# Plot the geometry
plot(sf::st_geometry(sc_strata))
}
}
\keyword{{datasets}}
