% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scallop_strata.R
\docType{data}
\name{scallop_strata}
\alias{scallop_strata}
\title{Sea Scallop Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -74.9 ymin: 36.6 xmax: -66.3 ymax: 42.2}
\item{Projected CRS}{WGS 84}
\item{Features}{23 x 8}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('scallop_strata')
}
\description{
An \code{sf} object containing spatial data for the Sea Scallop Survey. The current (as of 2025) NEFSC Scallop Strata (SAMS). Starting in 2025, a spatially balanced form of stratified random sampling called \if{html}{\out{<a href="https://www.nefmc.org/calendar/aug-19-2024-ssc-sub-panel-meeting">}}Generalized Random Tessellation Stratified (GRTS) was implemented, along with a switch to using Scallop Area Management Simulator (SAMS)\if{html}{\out{</a>}} areas as survey strata. While still a form of stratified random sampling, this method prevents random stations from being placed very close together. Use of the SAMS areas as strata ensures that sampling is more useful to management.\if{html}{\out{</p>}}
}
\examples{
\dontrun{
# Load the dataset
data(scallop_strata)

# Plot the geometry
plot(sf::st_geometry(scallop_strata))
}
}
\keyword{{datasets}}
