% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scallop_strata_historic.R
\docType{data}
\name{scallop_strata_historic}
\alias{scallop_strata_historic}
\title{Sea Scallop Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -76 ymin: 35.5 xmax: -65.9 ymax: 42.2}
\item{Projected CRS}{WGS 84}
\item{Features}{97 x 13}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('scallop_strata_historic')
}
\description{
An \code{sf} object containing spatial data for the Sea Scallop Survey. The historic (pre-2025) NEFSC Shellfish Strata. The standardized scallop dredge survey has been conducted since 1975 (annually since 1977) and employs a stratified random sampling design. The original NEFSC Shellfish Strata were based primarily on depth and secondarily by latitude, and locations of sampling stations were randomly selected within each stratum prior to each annual cruise.
}
\examples{
\dontrun{
# Load the dataset
data(scallop_strata_historic)

# Plot the geometry
plot(sf::st_geometry(scallop_strata_historic))
}
}
\keyword{{datasets}}
