% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-seal_stations.R
\docType{data}
\name{seal_stations}
\alias{seal_stations}
\title{Seal Aerial Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POINT}
\item{Dimension}{NA}
\item{Bounding box}{xmin: -70.9 ymin: 41.3 xmax: -67 ymax: 44.9}
\item{Projected CRS}{WGS 84}
\item{Features}{1026 x 14}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('seal_stations')
}
\description{
An \code{sf} object containing spatial data for the Seal Aerial Survey.
}
\examples{
\dontrun{
# Load the dataset
data(seal_stations)

# Plot the geometry
plot(sf::st_geometry(seal_stations))
}
}
\keyword{{datasets}}
