% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stock_footprints.R
\docType{data}
\name{stock_footprints}
\alias{stock_footprints}
\title{Data defining stock footprints}
\format{
A data frame/tibble
\describe{
\item{common_name}{Common name of species}
\item{scientific_name}{Scientific name of species}
\item{itis}{Species ITIS code}
\item{stock}{Stock definition for species, eg. Gulf of Maine, Georges Bank, etc.}
\item{strata}{The bottom trawl survey strata that the stock is found in}
}
}
\description{
A data frame containing stock footprints based on the bottom trawl survey strata
}
\keyword{datasets}
