% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-turtle_strata.R
\docType{data}
\name{turtle_strata}
\alias{turtle_strata}
\title{Turtle Ecology Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -82.1 ymin: 24.2 xmax: -61.8 ymax: 45.2}
\item{Projected CRS}{WGS 84}
\item{Features}{1 x 5}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('turtle_strata')
}
\description{
An \code{sf} object containing spatial data for the Turtle Ecology Survey.
}
\examples{
\dontrun{
# Load the dataset
data(turtle_strata)

# Plot the geometry
plot(sf::st_geometry(turtle_strata))
}
}
\keyword{{datasets}}
