% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_amapps_grid.R
\docType{data}
\name{amapps_grid}
\alias{amapps_grid}
\title{AMAPPS 10 km Grid}
\format{
A \code{raster} object:
\describe{
\item{class}{RasterLayer}
\item{dimensions}{328, 240, 78720 (nrow, ncol, ncell)}
\item{resolution}{10000 m, 10000 m (x, y)}
\item{extent}{-820443.2, 1579557, -1234728, 2045272 (xmin, xmax, ymin, ymax)}
\item{crs}{+proj=omerc +lat_0=35 +lonc=-75 +alpha=40 +gamma=40 +k=0.9996 +x_0=0 +y_0=0 +ellps=GRS80 +units=m +no_defs}
\item{source}{data-raw/gis/AMAPPS_10km_grid.tif}
}
}
\source{
\url{https://www.fisheries.noaa.gov/new-england-mid-atlantic/population-assessments/atlantic-marine-assessment-program-protected}
}
\usage{
data("amapps_grid")
}
\description{
A \code{raster} object containing the Atlantic Marine Assessment Program for Protected Species (AMAPPS) 10 km grid.
The AMAPPS grid has a cell size of 10 km x 10 km, with an oblique Mercator projection.
Grid cell values represent grid cell IDs.
}
\keyword{datasets}
