% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_BTS_strata.R
\docType{data}
\name{BTS_Strata}
\alias{BTS_Strata}
\title{NEFSC Bottom Trawl Survey Strata}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{dimension}{XY}
\item{Bounding box}{xmin: -79.18422 ymin: 32.50221 xmax: -65.16869 ymax: 44.83202}
\item{Geodetic CRS}{NAD 83}
\item{Features}{185 x 9}
\item{Metadata}{See source}
}
}
\source{
\url{https://drive.google.com/drive/folders/15s70lDwezatlNVReDMWRJadtlklMYSnC}

\url{https://www.fisheries.noaa.gov/inport/item/33471}
}
\usage{
BTS_Strata
}
\description{
An \code{sf} object containing the boundaries of the NEFSC bottom trawl survey strata.
This set of strata is considered historical. (Pre 2000).
}
\keyword{datasets}
