% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_EcoMon_Strata.R
\docType{data}
\name{EcoMon_Strata}
\alias{EcoMon_Strata}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\format{
A data frame/ \code{sf} object
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -75.96791 ymin: 35.14219 xmax: -65.16869 ymax: 44.48558}
\item{Geodetic CRS}{NAD83}
\item{Features}{47 x 4}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.fisheries.noaa.gov/inport/item/33468}
}
\usage{
data("EcoMon_Strata")
}
\description{
Ecosystem Monitoring Cruise (EcoMon) strata
}
\details{
An sf object containing the EcoMon strata
}
\note{
Use \url{https://mdeb-nefsc-noaa.hub.arcgis.com/datasets/noaa::ecosystem-monitoring-survey/about} instead.
}
\keyword{datasets}
