% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mdeb_datasets.R
\name{find_mdeb_datasets}
\alias{find_mdeb_datasets}
\title{Get names of data sets from MDEB database}
\usage{
find_mdeb_datasets()
}
\value{
data.frame
\item{title}{Name of the dataset}
\item{name}{Name of the dataset layer}
\item{id}{ID of the dataset layer}
\item{url}{URL of the dataset}
}
\description{
This function retrieves the names of datasets available in the
\href{https://mdeb-nefsc-noaa.hub.arcgis.com/}{MDEB} (Marine Data and
Environmental Biology) database.
}
