% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdeb_data.R
\name{get_mdeb_data}
\alias{get_mdeb_data}
\title{Get data from the MDEB (Marine Development and Ecology Branch) GIS Data Hub
hosted on NOAA's Geoplatform}
\usage{
get_mdeb_data(dataset_name = NULL, dataset_id = NULL)
}
\arguments{
\item{dataset_name}{Name of the dataset to retrieve.}

\item{dataset_id}{ID of the dataset layer to retrieve.}
}
\value{
An sf data.frame-like object
}
\description{
This function retrieves data from the \href{https://mdeb-nefsc-noaa.hub.arcgis.com/}{MDEB GIS Data Hub}.
}
\section{Available data}{

The list of available data and corresponding REST URLs
can be found by running \code{find_mdeb_data()}.
}

\examples{
# Pull EcoMon Survey strata from MDEB GIS Data Hub
ecomon_strata <- get_mdeb_data(dataset_name = "Ecosystem Monitoring Survey", id = 1)
# Plot the data
plot(ecomon_strata)

}
