% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdeb_datasets.R
\name{get_mdeb_datasets}
\alias{get_mdeb_datasets}
\title{Get sf data sets from MDEB database}
\usage{
get_mdeb_datasets(dataset_name = NULL, id = NULL)
}
\arguments{
\item{dataset_name}{Name of the dataset to retrieve.}

\item{id}{ID of the dataset layer to retrieve.}
}
\value{
data.frame
}
\description{
This function retrieves the names of datasets available in the
\href{https://mdeb-nefsc-noaa.hub.arcgis.com/}{MDEB} (Marine Data and
Environmental Biology) database.
}
\section{Data set names}{

The suite of available data set names and associated id's are found by
running \code{find_mdeb_datasets()}.
}

