% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_GillnetClosureAreas_20211021.R
\docType{data}
\name{GillnetClosureAreas_20211021}
\alias{GillnetClosureAreas_20211021}
\title{Gillnet Closure Areas}
\format{
An \code{sf} object
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -70.33333 ymin: 40.33333 xmax: -68.5 ymax: 41.5}
\item{Geodetic CRS}{NAD83}
\item{Features}{2 x 26}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.fisheries.noaa.gov/resource/map/gillnet-closure-areas}
}
\usage{
data("GillnetClosureAreas_20211021")
}
\description{
An \code{sf} object containing the Closure areas for Gillnets
Specifically Closed Area 1 and Nantucket Lightship Closure Area
}
\examples{
\dontrun{
# Load the dataset
data(GillnetClosureAreas_20211021)

# Plot the geometry
plot(sf::st_geometry(GillnetClosureAreas_20211021))
}
}
\keyword{datasets}
