% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data_Great_South_Channel_Restricted_Trap_Pot_Area.R
\docType{data}
\name{Great_South_Channel_Restricted_Trap_Pot_Area}
\alias{Great_South_Channel_Restricted_Trap_Pot_Area}
\title{Great South Channel Restricted Area}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -69.75 ymin: 41 xmax: -68.21667 ymax: 42.16667}
\item{Geodetic CRS}{NAD83}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.fisheries.noaa.gov/resource/map/great-south-channel-restricted-area}
}
\usage{
data("Great_South_Channel_Restricted_Trap_Pot_Area")
}
\description{
An \code{sf} object that depicts the boundaries of the Great South Channel Restricted Area in effect from April 1 to June 30 annually for the NOAA Fisheries Service’s Greater Atlantic Regional Fisheries Office (GARFO).
}
\keyword{datasets}
