% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Herring_Management_Areas.R
\docType{data}
\name{Herring_Management_Areas}
\alias{Herring_Management_Areas}
\title{Herring Management Areas}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{MULTIPOLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -78.57412 ymin: 33.81288 xmax: -65.69947 ymax: 45.1918}
\item{Geodetic CRS}{NAD83}
\item{Features}{4 x 25}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.fisheries.noaa.gov/resource/map/herring-management-areas}
}
\usage{
data("Herring_MAnagement_Areas")
}
\description{
An \code{sf} object depicting the boundaries of the Herring Management Areas
}
\keyword{datasets}
