% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_LMA1_Restricted_Area_FR.R
\docType{data}
\name{LMA1_Restricted_Area_FR}
\alias{LMA1_Restricted_Area_FR}
\title{LMA 1 Restricted Area}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -69.6128 ymin: 42.89198 xmax: -68.28787 ymax: 43.73333}
\item{Geodetic CRS}{WGS 84}
\item{Features}{1 x 25}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.fisheries.noaa.gov/resource/map/lma-1-restricted-area}
}
\usage{
data("LMA1_Restricted_Area_FR")
}
\description{
An \code{sf} object that depicts the boundaries of the Lobster Management Area One Restricted Area in effect annually from October 1 to January 30 for the NOAA Fisheries Service’s Greater Atlantic Regional Fisheries Office (GARFO).
}
\keyword{datasets}
