% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_NAFO_Divisions_2021_line_not_enclosed.R
\docType{data}
\name{NAFO_Divisions_2021_line_not_enclosed}
\alias{NAFO_Divisions_2021_line_not_enclosed}
\title{Northwest Atlantic Fisheries Organization (NAFO) 2021 Divisions}
\format{
A data frame/ sf object (42 x 7)
\describe{
\item{Label}{Label at SubSubDiv level}
\item{SubArea}{Area codes}
\item{Division}{Division Codes}
\item{SubDiv}{Codes for nested Divisions within Division}
\item{SubSubDiv}{Codes for nested Divisions within SubDiv}
\item{X}{x coordinate of polygon center}
\item{y}{y coordinate of polygon center}
}
}
\source{
\url{https://www.nafo.int/Data/GIS}
}
\usage{
NAFO_Divisions_2021_line_not_enclosed
}
\description{
An sf object containing the NAFO Divisions (LINESTRING geometry)
Pulled from https://www.nafo.int/Data/GIS
}
\section{Projection}{


CRS: NAD83
}

\seealso{
Other NAFO: 
\code{\link{FootprintAreaProjected}},
\code{\link{NAFO_Divisions_2021_line_enclosed}},
\code{\link{NAFO_Divisions_2021_poly_clipped}},
\code{\link{NAFO_Divisions_2021_poly_not_clipped}},
\code{\link{NAFO_Divisions_2021_pts}}
}
\concept{NAFO}
\keyword{datasets}
