% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Shellfish_strata.R
\docType{data}
\name{Shellfish_Strata}
\alias{Shellfish_Strata}
\title{Shellfish strata (Clam and Scallop surveys)}
\format{
A data frame/ \code{sf} object
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -75.95982 ymin: 35.50142 xmax: -65.89944 ymax: 42.16817}
\item{Geodetic CRS}{NAD 83}
\item{Metadata}{See source}
\item{Features}{97 x 9}
}
}
\source{
\url{https://www.fisheries.noaa.gov/inport/item/33466}
}
\usage{
Shellfish_Strata
}
\description{
An sf object containing the Clam and Scallop survey strata
}
\keyword{datasets}
