% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shrimp_strata.R
\docType{data}
\name{shrimp_strata}
\alias{shrimp_strata}
\title{Northern Shrimp Survey}
\format{
An \code{sf} object:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -70.7 ymin: 41.1 xmax: -68 ymax: 44.1}
\item{Projected CRS}{WGS 84}
\item{Features}{12 x 9}
\item{Metadata}{See source}
}
}
\source{
{https://mdeb-nefsc-noaa.hub.arcgis.com/}
}
\usage{
data('shrimp_strata')
}
\description{
An \code{sf} object containing spatial data for the Northern Shrimp Survey. Scientific survey strata for the Northern Shrimp Survey.
}
\examples{
\dontrun{
# Load the dataset
data(shrimp_strata)

# Plot the geometry
plot(sf::st_geometry(shrimp_strata))
}
}
\keyword{{datasets}}
