% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Shrimp_strata.R
\docType{data}
\name{Shrimp_Strata}
\alias{Shrimp_Strata}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\format{
A data frame/ \code{sf} object
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -70.68152 ymin: 41.08476 xmax: -67.99941 ymax: 44.08762}
\item{Geodetic CRS}{NAD83}
\item{Features}{12 x 5}
\item{Metadata}{See source}
}
}
\source{
\url{https://www.fisheries.noaa.gov/inport/item/33470}
}
\usage{
data("Shrimp_Strata")
}
\description{
Shrimp survey strata
}
\details{
An sf object containing the Shrimp survey strata
}
\note{
Use \url{https://mdeb-nefsc-noaa.hub.arcgis.com/datasets/noaa::northern-shrimp-survey/about} instead.
}
\keyword{datasets}
