% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Statistical_Areas_2010.R
\docType{data}
\name{Statistical_Areas_2010}
\alias{Statistical_Areas_2010}
\title{Greater Atlantic Region Statistical Areas}
\format{
A data frame/ sf object (135 x 1)
\describe{
\item{Id}{Area Number}
}
}
\source{
\url{https://www.fisheries.noaa.gov/resource/map/greater-atlantic-region-statistical-areas}
}
\usage{
Statistical_Areas_2010
}
\description{
An sf object containing the Greater Atlantic Region Statistical Areas.
Pulled from ftp://ftp.nefsc.noaa.gov/pub/gis/
}
\keyword{datasets}
