% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Statistical_Areas_2010_withNames.R
\docType{data}
\name{Statistical_Areas_2010_withNames}
\alias{Statistical_Areas_2010_withNames}
\title{Greater Atlantic Region Statistical Areas}
\format{
A data frame/ sf object (135 x 6)
\describe{
\item{Id}{Area Number}
\item{NAFODIV}{NAFO Division code}
\item{FULL_NAME}{Description of Division area}
\item{SHORT_NAME}{Abbreviated version of FULL_NAME}
\item{X}{x coordinate of polygon center}
\item{y}{y coordinate of polygon center}
}
}
\source{
\url{https://www.fisheries.noaa.gov/resource/map/greater-atlantic-region-statistical-areas}
}
\usage{
Statistical_Areas_2010_withNames
}
\description{
An sf object containing the Greater Atlantic Region Statistical Areas.
Pulled from ftp://ftp.nefsc.noaa.gov/pub/gis/
}
\keyword{datasets}
